+++ src/minitests/mod.rs
@@ -283,22 +283,6 @@ pub fn test_crosscontract_call_with_constructors() {
 }
 
 #[test]
-pub fn test_gas_charging_underfunded() {
-    match crate::evm::_evm_run_with_gas_charging(None, Uint256::_from_gwei(20), false, false) {
-        Ok(result) => assert_eq!(result, false),
-        Err(e) => panic!("error {}", e),
-    }
-}
-
-#[test]
-pub fn test_gas_charging_fully_funded() {
-    match crate::evm::_evm_run_with_gas_charging(None, Uint256::_from_eth(1000), false, false) {
-        Ok(result) => assert_eq!(result, true),
-        Err(e) => panic!("error {}", e),
-    }
-}
-
-#[test]
 pub fn test_tx_with_deposit() {
     match crate::evm::_evm_tx_with_deposit(None, false, false) {
         Ok(result) => assert_eq!(result, true),
