+++ src/upload.rs
@@ -10,7 +10,6 @@ use std::fmt;
 use std::fs::File;
 use std::io::Read;
 use std::path::Path;
-use crate::compile::miniconstants::ARBOS_VERSION;
 
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub struct CodeUploader {
@@ -166,61 +165,4 @@ fn test_code_upload_prep() {
     assert_eq!(uploader.num_total, uploader.num_so_far);
     let reconstituted = CodeUploader::_from_json(&uploader.clone()._to_json().unwrap());
     assert_eq!(uploader, reconstituted);
-}
-
-#[test]
-fn _test_upgrade_arbos_to_different_version() {
-    _test_upgrade_arbos_over_itself_impl().unwrap();
-}
-
-fn _test_upgrade_arbos_over_itself_impl() -> Result<(), ethabi::Error> {
-    let mut machine = load_from_file(Path::new("arb_os/arbos_before.mexe"));
-    machine.start_at_zero();
-
-    let wallet = machine.runtime_env.new_wallet();
-    let my_addr = Uint256::from_bytes(wallet.address().as_bytes());
-
-    let mut add_contract = AbiForContract::new_from_file(&test_contract_path("Add"))?;
-    if add_contract
-        .deploy(&[], &mut machine, Uint256::zero(), None, false)
-        .is_err()
-    {
-        panic!("failed to deploy Add contract");
-    }
-
-    let arbowner = _ArbOwner::_new(&wallet, false);
-
-    let arbsys_orig_binding = ArbSys::new(&wallet, false);
-    assert_eq!(
-        arbsys_orig_binding._arbos_version(&mut machine)?,
-        Uint256::one()
-    );
-
-    arbowner._give_ownership(&mut machine, my_addr, Some(Uint256::zero()))?;
-
-    let uploader = CodeUploader::_new_from_file(Path::new("arb_os/arbos-upgrade.mexe"));
-    arbowner._start_code_upload(&mut machine)?;
-
-    let mut accum = vec![];
-    for buf in uploader.instructions {
-        accum.extend(buf);
-        if (accum.len() > 3000) {
-            arbowner._continue_code_upload(&mut machine, accum)?;
-            accum = vec![];
-        }
-    }
-    if (accum.len() > 0) {
-        arbowner._continue_code_upload(&mut machine, accum)?;
-    }
-
-    arbowner._finish_code_upload_as_arbos_upgrade(&mut machine)?;
-
-    let wallet2 = machine.runtime_env.new_wallet();
-    let arbsys = ArbSys::new(&wallet2, false);
-    let arbos_version = arbsys._arbos_version(&mut machine)?;
-    assert_eq!(arbos_version, Uint256::from_u64(ARBOS_VERSION));
-    let arbos_version_orig = arbsys_orig_binding._arbos_version(&mut machine)?;
-    assert_eq!(arbos_version, arbos_version_orig);
-
-    Ok(())
 }
